[SW-2079] Move api classes to ai.h2o.sparkling package
