+++ core/src/main/resources/META-INF/services/water.api.Schema
@@ -1,11 +1,11 @@
-water.api.DataFrames.DataFrameV3
-water.api.DataFrames.H2OFrameIDV3
-water.api.DataFrames.DataFramesV3
-water.api.H2OFrames.DataFrameIDV3
-water.api.RDDs.RDD2H2OFrameIDV3
-water.api.RDDs.RDDV3
-water.api.RDDs.RDDsV3
-water.api.scalaInt.ScalaCodeV3
-water.api.scalaInt.ScalaSessionIdV3
-water.api.scalaInt.ScalaSessionsV3
-water.api.scalaInt.ScalaCodeResultV3
+ai.h2o.sparkling.backend.api.dataframes.DataFrameV3
+ai.h2o.sparkling.backend.api.dataframes.H2OFrameIDV3
+ai.h2o.sparkling.backend.api.dataframes.DataFramesV3
+ai.h2o.sparkling.backend.api.h2oframes.DataFrameIDV3
+ai.h2o.sparkling.backend.api.rdds.RDD2H2OFrameIDV3
+ai.h2o.sparkling.backend.api.rdds.RDDV3
+ai.h2o.sparkling.backend.api.rdds.RDDsV3
+ai.h2o.sparkling.backend.api.scalainterpreter.ScalaCodeV3
+ai.h2o.sparkling.backend.api.scalainterpreter.ScalaSessionIdV3
+ai.h2o.sparkling.backend.api.scalainterpreter.ScalaSessionsV3
+ai.h2o.sparkling.backend.api.scalainterpreter.ScalaCodeResultV3
