+++ core/src/main/scala/ai/h2o/sparkling/backend/api/dataframes/DataFrameV3.java
@@ -1,4 +1,4 @@
-package water.api.DataFrames;
+package ai.h2o.sparkling.backend.api.dataframes;
 
 
 import water.api.API;
