+++ core/src/main/scala/ai/h2o/sparkling/backend/api/scalainterpreter/ScalaCodeResult.java
@@ -1,4 +1,4 @@
-package water.api.scalaInt;
+package ai.h2o.sparkling.backend.api.scalainterpreter;
 
 import water.Key;
 import water.Lockable;
