+++ core/src/main/scala/ai/h2o/sparkling/backend/external/ExternalBackendUtils.scala
@@ -19,11 +19,11 @@ package ai.h2o.sparkling.backend.external
 
 import java.net.{InetAddress, NetworkInterface}
 
+import ai.h2o.sparkling.backend.api.RestAPIManager
 import ai.h2o.sparkling.backend.utils.SharedBackendUtils
 import org.apache.spark.SparkEnv
 import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.h2o.{H2OConf, H2OContext}
-import water.api.RestAPIManager
 import water.init.{HostnameGuesser, NetworkBridge}
 import water.{H2O, H2OStarter, Paxos}
